Move `find_codec_by_name` function into the `codecs` package module
