+++ multiaddr/codecs/fspath.py
@@ -3,7 +3,7 @@ import os
 
 import six
 
-from ..codec import LENGTH_PREFIXED_VAR_SIZE
+from . import LENGTH_PREFIXED_VAR_SIZE
 
 
 SIZE = LENGTH_PREFIXED_VAR_SIZE
