+++ multiaddr/codecs/p2p.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 import base58
 import six
 
-from ..codec import LENGTH_PREFIXED_VAR_SIZE
+from . import LENGTH_PREFIXED_VAR_SIZE
 
 
 SIZE = LENGTH_PREFIXED_VAR_SIZE
